Added bareostools
bareostools consists of bacula disaster recovery tools such as

bls, bscan, and bextract

These are packages that need to be installed to try these things out.

In production though you may not want to leave them lying around your
server as it makes it pretty easy to extract your data, so it's probably
better to keep these packages on your local machine and connect to them
with  sshfs.