BareOS Puppet Module Modified JobDef selection method
When working with Backup Copy Jobs I noticed that the copy job I was
using kept getting all the attributes of the JobDef it was being
assigned to.

Honestly it doesn't need any of the attributes of the default job for a
Backup Copy Job and so I changed the way in which the
bareos::director::job class determines if it is a JobDef or not.

I also changed the way the template determines if a parent JobDef should
be specified at all in the resulting configuration file,
and if the type is Copy the parent JobDef will not be specified at all.