Added GFS19 of a single webserver job and testing
THIS IS AN IMPORTANT COMMIT

I added a job (webserver-job) with 3 different pools for volumes:

- webserver-monthly-pool
(Volumes end up in /mnt/backups/webserver/monthly)

- webserver-weekly-pool
(Volumes ends up in /mnt/backups/webserver/weekly)

- webserver-daily-pool
(Volumes end up in /mnt/backups/webserver/daily)

Actually new clients (fd) with the same format of GFS can be easily
added by editing:

- modules/bareosdir/manifests/init.pp

- modules/bareosdir/manifests/common/schedule.pp

Also in this commit is included a testing framework for testing the
backups in an iterative fashion by changing the dates of the VMs in
VirtualBox; it also works well for demonstration purposes.

It is contained in the file ./updateTimesNWeeks.sh which you can see
along with README.MD for more details.