Copy-Jobs work but still need fixed
The Copy-Jobs are being successfully copied to the storage daemon, but
the schedules aren't set to copy them when they're supposed to be copied
during the day.

Also the time on the storage daemon needs to be updated to match that of
the director.  NTP would be the real way to make this work.